Fix colors and styling for commit list and field changes.
When working on the diff theme support, the commit lists ended up with
a broken background color in light mode for the "-" and "+" operators.
This happened because a variable these depended on became an alpha
overlay color, rather than using a base color for insert/delete.

All the diff op colors have been updated to use base colors, ensuring
consistency there in both light and dark modes. The commits list has
also been updated to use a more correct font size in the process, so the
summary and message don't appear substantially larger than the
surrounding elements.

The standard file fields have been updated to match the same colors as
well, along with the same icons shown in the commit list and the same
padding. This helps bring consistency between the two elements. Right
now, these definitions are duplicated. Work needs to be done to build
CSS components for this part of the UI.

Testing Done:
Tested several fields and the commit list in both light and dark modes.

Reviewed at https://reviews.reviewboard.org/r/13902/