Update rb-c-alert to use ink-c-alert's visuals.
We have code that uses `rb-c-alert`, and will continue to through
release. To help keep things looking consistent, this change updates the
colors and the icons to match the visuals of `ink-c-alert`.

Testing Done:
Tested all alert types in light and dark mode (initially as part of the
diff file notice change, before a new approach was taken), and verified
that the right icons and colors were used.

Reviewed at https://reviews.reviewboard.org/r/13903/