Add more default columns to the base review request datagrid.
This adds the Ship It/Issue counts and New Updates columns to the defaults
for the base review request datagrid, which is used for the All Review Requests
page. I debated adding the My Comments column too but decided against it.
I felt that it was too personal and that the All Review Requests page should
default to just general information instead.

I also audited the columns we have for the dashboard.

Testing Done:
- Saw the new columns added to the All Review Requests page (while logged in
and logged out)
- Saw that the columns on the dashboard appeared as normal
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13909/