Update the Review Board logo URL that we suggest to upload to MS Teams in the docs.
When setting up a Microsoft Teams integration, admins need to create an
incoming WebHook via the Microsoft Teams app. When creating the WebHook, you
have the ability to upload a picture to associate with all messages coming
from the WebHook (basically an avatar).

Since we don't have control over this, we offer a link to a Review Board logo
in our docs that admins can upload. This logo differs from our normal logo by
being centered, so that it looks nice in the rounded crop that most chat
applications use for avatars. We recently added this centered logo to our
CDN, so this updates the docs to use the right URL for the centered logo.

Testing Done:
Accessed the URL and made sure it lead to the centered logo.

Reviewed at https://reviews.reviewboard.org/r/13916/