Don't trigger a second click on action menu items when clicked.
When converting `MenuActionView` over to `Ink.MenuView`, the `onClick`
handler was set up to invoke an explicit click on the item element if an
`activate` method wasn't defined by the handler.

This impacted link-based menu items. While clicking still worked, we
technically ended up invoking the click twice. Once via the browser
(which gave us the ability to Control/Cmd-Click to open in a new tab),
and once by via the explicit click trigger. So, if trying to open in a
new tab, we'd succesfully open in a new tab but also open the same link
in the current tab.

We no longer provide this explicit action. Best I can tell, we don't
need that explicit click on the element anywhere.

Testing Done:
Tested the Follow menu, Account menu, Review menu, and all the review
request action menus. Didn't see any regressions in behavior.

Tested Cmd-clicking Follow menu items and Account menu items. Saw that
they opened in new tabs, but not in the current tab.

Reviewed at https://reviews.reviewboard.org/r/13918/