Fix spacing and alignment issues within the search fields.
The Search field at the top of the page and the Filter field on the New
Review Request page both had their own visual alignment and spacing
issues, in part due to altered styling in Ink and bad browser defaults.

The styling for `rb-c-search-field` now uses a more appropriate font
size for the FontAwesome-based icon, and a more appropriate spacing
between it and the text.

The issues in the Filter field were due to the lack of a `type=`
attribute, and the lack of default styling for an `<input>` without
`type=`. As fixing this globally for all type-less inputs would alter
precedence rules for other styles, I've opted not to touch those for
now. Instead, we now set the Filter field to `type="search"` explicitly.

Testing Done:
Tested in light and dark modes for both of these fields.

Reviewed at https://reviews.reviewboard.org/r/13920/