Fix scrolling in the text editor in the comment dialog.
The fix to prevent accidentally scrolling away from the comment dialog
broke scrolling within the text editor. This change adds some smarts so
we figure out if we're trying to scroll in the text editor and the text
editor is capable of scrolling, we'll allow it, and if not, we'll
continue to swallow the event.

Testing Done:
- Verified that the scroll wheel did not do anything when moused over
  parts of the comment dialog other than the text editor.
- Verified that the scroll wheel did not do anything when moused over an
  empty text field, in both markdown and plain text modes.
- Verified that the scroll wheel did not do anything when moused over
  text that did not overflow the text editor in markdown mode.
- Verified that the scroll wheel scrolled the text editor when working
  with long comments in both markdown and plan text modes.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/13929/