Update Review Board packaging to use Ink and Tabler Icons.
The `package-lock.json` now reflects the Ink dependency from Djblets,
and locks the version for building Review Board.

We make use of Tabler icons for the Follow menu, so we now need to
include this as a dependency as well.

Testing Done:
Successfully built packages and deployed.

Reviewed at https://reviews.reviewboard.org/r/13950/