Allow passing SCMTool IDs to the tool= parameter to the repositories API.
The `tool=` parameter for the repositories API was implemented using
SCMTool names, but we've had various issues in the past where names were
fluid for one reason or another. This change makes it so we can pass in
the SCMTool ID for this parameter as well.

Additionally, we had a bug where passing as unknown name to this
parameter would cause a crash in the API because the tool returned from
the registry lookup was `None`. This has been fixed to just skip adding
that to the list.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13972/