Use short buttons for the API token inline editors.
The buttons for the description inline editor for API tokens on the My
Account page were overlapping with the "Full access v" drop-down and the
"Remove" button. This change makes it so we use short buttons for the
description and expiration editors.

Testing Done:
Opened the editors for API tokens in the My Account authentication page.
Saw that they no longer overlapped with the token controls.

Reviewed at https://reviews.reviewboard.org/r/13983/