Make sure region comment coordinates are rounded to integers.
When saving the x, y, width and height coordinates for a region comment,
we previously would not round these and they would sometimes include
many decimal places.

This caused a bug with serializing comments that are made on the same
region (e.g. creating a new comment on top of an existing one). We represent
these types of comments as a single comment box that shows the number of
comments in the corner. When serializing the comments, we use a key
built of the x, y, width and height so that comments on the same region
get mapped to the same box.

Without rounding the decimal places for coordinates, there could be slight
differences in the coordinate values for the same visual region.
Comments created on top of one another would sometimes fail to be
serialized into the same comment box due to the decimal differences.

This change ensures that the coordinate values are rounded.

Testing Done:
Tested creating region comments on top of one another, saw that they
were properly placed into the same box instead of two separate
overlapping boxes.

Reviewed at https://reviews.reviewboard.org/r/13990/