Fix a crash when computing diff headers for filtered-out lines.
When attempting to generate diff headers on interdiffs, we can run into
a crash when processing this on a block of filtered-out lines. This is a
symptom of a larger issue with the interdiff filtering logic, which will
be dealt with separately, but can be easily worked around by checking
what we're comparing against before making assumptions about the data.

Testing Done:
Unit tests passed.

Verified that this fixed the crash occurring in an interdiff.

Reviewed at https://reviews.reviewboard.org/r/14001/