Work around a breaking LessCSS change in Djblets 5.1.
Djblets 5.1 made some changes to the common LessCSS variables, turning
some into CSS variables. This broke Review Board, which was trying to
apply preprocessor calculations to these variables.

We'll need to address this in a Djblets 5.1.x release, but for now,
we're working around this by removing the usage of one of these
variables and hard-coding a value instead.