Don't include draft diffsets in the initial latest timestamp.
When we load the diff viewer, we include the latest activity timestamp
in the page data. This is then used with subsequent accesses to the
last-update resource to determine whether to show the updates bubble.

The timestamp of a draft diffset was being included in the initial
activity timestamp, while it is not included in the timestamp returned
by the API. This meant that the updates bubble would pop up saying the
diff was new because the timestamp did not match.

Testing Done:
Loaded the diff viewer with a draft diff (while viewing as an admin
user). Saw that the updates bubble no longer popped up incorrectly.

Reviewed at https://reviews.reviewboard.org/r/14009/