Capture errors attempting to call draft.ready() from the unified banner.
The unified banner attempts to access the draft via the API for the
draft when its created, in order to start up its draft modes. When the
user does not have access to the draft, this ends up with an error in
the JS console. This error was harmless, but annoying to see when
debugging.

This change adds a catch so we can swalllow any errors when attempting
to call ready on the draft.

Testing Done:
Loaded a review request owned by another user and saw that I no longer
got an error in my devtools.

Reviewed at https://reviews.reviewboard.org/r/14012/