Tweak GitHub form help text a bit.
The help text for the GitHub access token was sticking a <strong> next
to a <code>, which ended up collapsing the space in between them. This
change simplifies the text a bit and gets rid of the <strong>, since it
doesn't really help and IMO just makes it harder to read.

This also adds some <code> tags to the example URLs for account/org/repo
names in the repository part of the form.

Testing Done:
Loaded the form for a GitHub repo and saw that the help text all looked
nice.

Reviewed at https://reviews.reviewboard.org/r/14026/