Prevent admins from editing fields when there's an un-viewed draft.
Now that admins can make a choice about whether they want to view an
unpublished draft or not, we need to be more careful about when they
make edits. This change makes it so if the admin attempts to make a
change to the review request when a user draft exists, we'll block the
edit and prompt them to load the draft first.

Testing Done:
- Attempted to edit review request fields, file attachment
  captions/state, and add new attachments/diffs as an admin while
  viewing a review request with an unpublished draft, but not viewing
  the draft. Saw the expected prompt.
- Edited review request fields, file attachment captions/state, and
  added new attachments/diffs as both the owner and an admin while
  viewing the current draft.

Reviewed at https://reviews.reviewboard.org/r/14037/