Activate enhanced typing for Django using django-stubs-ext.
`django-stubs` provides an extension for `mypy` that adds enhanced
typing for Django, including correct handling of model fields. This
change activates the plugin, giving us better typing by default.

We also now specify the version of `django-stubs` we want to depend on.

Testing Done:
Verified that the type checker was happy with several of the models
we have in the codebase.

Reviewed at https://reviews.reviewboard.org/r/14031/