Switch to draft mode when an admin edits a review request.
When an admin is viewing someone else's review request, they now have a
choice between viewing it as if they were a regular user, and seeing
unpublished draft data.

This change makes it so when a new draft is created by the admin, we
switch to the view-draft mode.

Testing Done:
Made changes to a review request (setting draft fields, creating
attachments) and saw that the draft banner appeared, the notice about
viewing an unpublished draft appeared, and that the URL in the browser
changed to include the ?view-draft=1 query parameter.

Reviewed at https://reviews.reviewboard.org/r/14038/