Make region comment flag and resize icon public.
This changes the flag and resize icon properties on the region comment
block view from private to public. Eventually when Power Pack drops
support for RB5 and 6, we'll need to be able to access these variables
via a child class of the region comment block view, which we can't do
if they're private.

This isn't a blocker for Power Pack 6. Since that needs to support
RB5 and 6, we override the flag and resize icon properties there
to suit our needs. Once Power Pack is on RB7+, we'll want to
inherit and use these public properties instead.

Testing Done:
- Ran JS unit tests.
- Smoked tested region comments on an image.

Reviewed at https://reviews.reviewboard.org/r/14055/